% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{createDummies}
\alias{createDummies}
\title{Deprecated: use \code{\link{create_dummies}} instead}
\usage{
createDummies(
  group.id,
  act.id,
  part.id,
  d,
  include.self = FALSE,
  merge.original = FALSE
)
}
\arguments{
\item{group.id}{string; name of the group identifier variable}

\item{act.id}{string; name of the actor identifier variable}

\item{part.id}{string; name of the partner identifier variable}

\item{d}{a \code{data.frame} in directed dyadic long-form}

\item{include.self}{logical; retain self-ratings? Default \code{FALSE}}

\item{merge.original}{logical; merge back onto original data? Default
\code{FALSE}}
}
\value{
see \code{\link{create_dummies}}
}
\description{
\code{createDummies} has been renamed to \code{create_dummies}.
The argument names have also changed: \code{group.id} -> \code{group_id},
\code{act.id} -> \code{act_id}, \code{part.id} -> \code{part_id},
\code{d} -> \code{data}, \code{include.self} -> \code{include_self},
\code{merge.original} -> \code{merge_original}.
}
