% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{srmRun}
\alias{srmRun}
\title{Deprecated: use \code{\link{srm_run}} instead}
\usage{
srmRun(dv, groupId, actId, partId, feVars = NULL, data)
}
\arguments{
\item{dv}{string; name of the outcome variable}

\item{groupId}{string; name of the group identifier variable}

\item{actId}{string; name of the actor identifier variable}

\item{partId}{string; name of the partner identifier variable}

\item{feVars}{character vector of fixed-effect predictor names, or
\code{NULL}}

\item{data}{a \code{data.frame} at the directed dyad level}
}
\value{
see \code{\link{srm_run}}
}
\description{
\code{srmRun} has been renamed to \code{srm_run}.
The argument names have also changed: \code{groupId} -> \code{group_id},
\code{actId} -> \code{act_id}, \code{partId} -> \code{part_id},
\code{feVars} -> \code{fe_vars}.
}
