% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.pdSRM.R
\name{coef.pdSRM}
\alias{coef.pdSRM}
\title{Extract model coefficients from pdSRM object}
\usage{
\method{coef}{pdSRM}(object, unconstrained = TRUE, ...)
}
\arguments{
\item{object}{an object inheriting from pdSRM}

\item{unconstrained}{a logical value. If TRUE the coefficients are returned
in unconstrained form. If FALSE the upper triangular elements are returned}

\item{...}{additional arguments}
}
\value{
a vector with the coefficients corresponding to object
}
\description{
This is used internally as part of the construction of the pdSRM
object that will fit the appropriate structure for the SRM.
}
\examples{
\dontrun{
o = lme(liking ~ 1, random=list(groupId=pdBlocked(list(pdIdent(~1),
pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
correlation=corCompSymm(form=~1 | groupId/pdSRM_dyad_id),
data=d, na.action=na.omit)
}
}
