% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrix.pdSRM.R
\name{corMatrix.pdSRM}
\alias{corMatrix.pdSRM}
\title{Extract Correlation Matrix from a pdSRM Object}
\usage{
\method{corMatrix}{pdSRM}(object, ...)
}
\arguments{
\item{object}{an object inheriting from pdSRM}

\item{...}{some methods for this require additional arguments}
}
\value{
the correlation matrix corresponding to the positive-definite
matrix represented by object
}
\description{
This function is used internally as part of the construction of the pdSRM
object that will fit the appropriate structure for the SRM. The correlation matrix
corresponding to the positive-definite matrix represented by object is
obtained.
}
\examples{
\dontrun{
o = lme(liking ~ 1, random=list(groupId=pdBlocked(list(pdIdent(~1),
pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
correlation=corCompSymm(form=~1 | groupId/pdSRM_dyad_id),
data=d, na.action=na.omit)
}
}
