% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDummies.R
\name{createDummies}
\alias{createDummies}
\title{Create dummy variables}
\usage{
createDummies(
  group.id,
  act.id,
  part.id,
  d,
  include.self = FALSE,
  merge.original = FALSE
)
}
\arguments{
\item{group.id}{string variable name that is the group identifier}

\item{act.id}{string variable name that is the actor identifier}

\item{part.id}{string variable name that is the partner identifier}

\item{d}{data.frame structured in directed dyadic long-form}

\item{include.self}{boolean indicating whether to include self ratings}

\item{merge.original}{boolean indicating whether to return the input
dataset with the new identifiers}
}
\value{
data.frame giving a new set of unique identifiers and the dummy
variables needed to run an SRM using multilevel modeling
}
\description{
Generate a series of dummy variables for the actor and partner, following
the process designed by Kenny & Snijders (1999). This generates N dummies
for actor and N dummies for partner, where N is the maximum group size
in the dataset
}
\examples{
o = createDummies(group.id="groupId", act.id="actId", part.id="partId", d=sampleDyadData)
}
