% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdMatrix.pdSRM.R
\name{pdMatrix.pdSRM}
\alias{pdMatrix.pdSRM}
\title{Extract Matrix or Square-Root Factor from a pdSRM object}
\usage{
\method{pdMatrix}{pdSRM}(object, factor = FALSE)
}
\arguments{
\item{object}{an object inheriting from pdSRM}

\item{factor}{an optional logical value}
}
\value{
if factor is FALSE the positive-definite matrix represented by object;
else a square-root of the positive-definite matrix is returned
}
\description{
This function is used internally as part of the construction of the pdSRM
object that will fit the appropriate structure for the SRM.
}
\examples{
\dontrun{
o = lme(liking ~ 1, random=list(groupId=pdBlocked(list(pdIdent(~1),
pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
correlation=corCompSymm(form=~1 | groupId/pdSRM_dyad_id),
data=d, na.action=na.omit)
}
}
