% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdSRM.R
\name{pdSRM}
\alias{pdSRM}
\title{Construct the pdSRM object}
\usage{
pdSRM(
  value = numeric(0),
  form = NULL,
  nam = NULL,
  data = sys.frame(sys.parent())
)
}
\arguments{
\item{value}{an optional initialization value}

\item{form}{an optional one-sided linear formula specifying the row/column names for the matrix}

\item{nam}{an optional vector of character strings specifying the row/column names for the matrix}

\item{data}{inherited from the surrounding nlme call}
}
\value{
a pdMat object representing a positive-definite matrix conforming to the SRM structure
}
\description{
This function is used to create the correct structure for the variance/covariance
matrix for a full social relations model.
}
\examples{
\dontrun{
o = lme(liking ~ 1, random=list(groupId=pdBlocked(list(pdIdent(~1),
pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
correlation=corCompSymm(form=~1 | groupId/pdSRM_dyad_id),
data=d, na.action=na.omit)
}
}
