% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsrm_run.R
\name{print.cpsrm}
\alias{print.cpsrm}
\title{Print a cpsrm Object}
\usage{
\method{print}{cpsrm}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{"cpsrm"}}

\item{digits}{integer; number of digits to print. Default \code{4}}

\item{...}{additional arguments (currently unused)}
}
\value{
\code{x}, invisibly
}
\description{
Print a cpsrm Object
}
