% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{srmPseudoRSq}
\alias{srmPseudoRSq}
\title{Deprecated: use \code{\link{srm_pseudo_rsq}} instead}
\usage{
srmPseudoRSq(null.model, predict.model)
}
\arguments{
\item{null.model}{an \code{lme} null model fitted with \code{\link{pdSRM}}}

\item{predict.model}{an \code{lme} predictor model fitted with
\code{\link{pdSRM}}}
}
\value{
see \code{\link{srm_pseudo_rsq}}
}
\description{
\code{srmPseudoRSq} has been renamed to \code{srm_pseudo_rsq}.
The argument names have also changed: \code{null.model} ->
\code{null_model}, \code{predict.model} -> \code{predict_model}.
}
