% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmPseudoRSq.R
\name{srmPseudoRSq}
\alias{srmPseudoRSq}
\title{Calculate a Pseudo R-Squared for the Social Relations Model}
\usage{
srmPseudoRSq(null.model, predict.model)
}
\arguments{
\item{null.model}{an lme output object fit using pdSRM, without covariates}

\item{predict.model}{an lme output object fit using pdSRM, with covariates}
}
\value{
data.frame with variances from a null and predict model and the
estimated pseudo r-squared values
}
\description{
Provides an estimated pseudo r-squared value based on the comparison
between a null model and a model with fixed effects
}
\examples{
o = srmPseudoRSq(null.model=nlme::lme(liking ~ 1,
random=list(groupId=nlme::pdBlocked(list(nlme::pdIdent(~1),
pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
correlation=nlme::corCompSymm(form=~1 | groupId/pdSRM_dyad_id),
data=
createDummies(group.id="groupId", act.id="actId", part.id="partId",
             d=sampleDyadData[sampleDyadData$timeId==1, ],
               merge.original=TRUE), na.action=na.omit),
               predict.model= nlme::lme(liking ~ groupCohesion +
               actEx + partEx + contact,
               random=list(groupId=nlme::pdBlocked(list(nlme::pdIdent(~1),
               pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
               correlation=nlme::corCompSymm(form=~1 |
               groupId/pdSRM_dyad_id),
             createDummies(group.id="groupId",
             act.id="actId", part.id="partId",
             d=sampleDyadData[sampleDyadData$timeId==1, ],
             merge.original=TRUE), na.action=na.omit))
}
