% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{srmVarPct}
\alias{srmVarPct}
\title{Deprecated: use \code{\link{srm_var_pct}} instead}
\usage{
srmVarPct(object)
}
\arguments{
\item{object}{an \code{lme} model object fitted with \code{\link{pdSRM}}}
}
\value{
see \code{\link{srm_var_pct}}
}
\description{
\code{srmVarPct} has been renamed to \code{srm_var_pct}.
}
