% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmVarPct.R
\name{srmVarPct}
\alias{srmVarPct}
\title{Extract Variance Decomposition Results from pdSRM Output}
\usage{
srmVarPct(object)
}
\arguments{
\item{object}{an lme output object with the SRM structure}
}
\value{
data.frame with the variance-covariance parameters, percentages, and
the reciprocity correlations
}
\description{
Returns the breakdown of variance components and reciprocity correlations
from the output of a multilevel model fitting the SRM according to the
dummy variable specification.
}
\examples{
o = srmVarPct(nlme::lme(liking ~ 1, random=list(groupId=nlme::pdBlocked(list(nlme::pdIdent(~1),
pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
correlation=nlme::corCompSymm(form=~1 | groupId/pdSRM_dyad_id),
data=
createDummies(group.id="groupId", act.id="actId", part.id="partId",
d=sampleDyadData[sampleDyadData$timeId==1, ],
merge.original=TRUE), na.action=na.omit))
}
