% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsrmRun.R
\name{summary.cpsrm}
\alias{summary.cpsrm}
\title{Summarize a cpsrm Object}
\usage{
\method{summary}{cpsrm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"cpsrm"}}

\item{...}{additional arguments passed to \code{\link{print.cpsrm}}}
}
\value{
\code{object}, invisibly
}
\description{
Summarize a cpsrm Object
}
