% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pdSRM.R
\name{summary.pdSRM}
\alias{summary.pdSRM}
\title{Summarize a pdSRM Object}
\usage{
\method{summary}{pdSRM}(object, structName = "Social Relations Model", ...)
}
\arguments{
\item{object}{an object inheriting from pdSRM}

\item{structName}{an optional character string with a description of the pdSRM class}

\item{...}{optional arguments for some methods}
}
\value{
an object similar to object, with additional attributes structName
}
\description{
This function is used internally as part of the construction of the pdSRM
object that will fit the appropriate structure for the SRM.
Attributes structNsame and noCorrelation, with the values of the
corresponding arguments to the method function, are appended
to object and its class is changed to summary.pdSRM
}
\examples{
\dontrun{
o = lme(liking ~ 1, random=list(groupId=pdBlocked(list(pdIdent(~1),
pdSRM(~-1 + a1 + a2 + a3 + a4 + p1 + p2 + p3 + p4)))),
correlation=corCompSymm(form=~1 | groupId/pdSRM_dyad_id),
data=d, na.action=na.omit)
}
}
